requirements.txt
langchain-mcp-adapters
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.1.0
langsmith>=0.1.0
requests>=2.28.0
pyyaml>=6.0
colorama>=0.4.0
python-dotenv>=1.0.0
